refactor(core): verify address moved from core to account lib
Verify address moved from core to account lib
